requirements.txt
asgiref==3.8.1
Django==5.0.3
django-cors-headers==4.3.1
django-jazzmin==2.6.1
djangorestframework==3.15.1
environs==11.0.0
marshmallow==3.21.1
packaging==24.0
pillow==10.3.0
python-dotenv==1.0.1
sqlparse==0.4.4
tzdata==2024.1
